Prepare for deployment (test/WIP)

+++ database/migrations/2024_10_02_165325_create_payroll_items_table.php
@@ -34,7 +34,7 @@ return new class extends Migration
         });
 
         Schema::create('additions', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigInteger('id')->unsigned();
             $table->primary('id');
             $table->string('name');
             $table->text('description');
@@ -57,7 +57,7 @@ return new class extends Migration
         });
 
         Schema::create('deductions', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigInteger('id')->unsigned();
             $table->primary('id');
             $table->string('name');
             $table->text('description');
